get rid of test utils, that was dumb
